% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapreg.R
\docType{data}
\name{mapregions}
\alias{mapregions}
\title{Maps visualization}
\usage{
regmap(geofile, territory, measure, ...)
}
\arguments{
\item{geofile}{geoJson file to read.}

\item{territory}{vector of characters/factor; territories field (usualy countries, cities, etc) to choose.}

\item{measure}{numeric (usually); in the moste cases a quantitative variable to be setted up as measure (for example: location area, etc)}

\item{label}{string; the name of measure field.}

\item{main}{string; the title of map.}

\item{mapcol}{color of specific regions.}

\item{xlimit}{numeric; x-axis limits.}

\item{ylimit}{numeric; y-axis limits.}
}
\value{
plot
}
\description{
Plot any territory given Geojson file containing coordinates.
}
